+++ integration_tests/suite/base/test_users.py
@@ -963,6 +963,7 @@ def generate_user_resources_bodies(
     device=None,
     user_destination=None,
     queue=None,
+    endpoint_name=None
 ):
     exten = h.extension.find_available_exten(context_name)
     vm_number = h.voicemail.find_available_number(context_name)
@@ -973,10 +974,13 @@ def generate_user_resources_bodies(
     user = FULL_USER
     auth = AUTH_USER
     extension = {'context': context_name, 'exten': exten}
+    if not endpoint_name:
+        endpoint_name = s.random_string(5)
+
     line = {
         'context': context_name,
         'extensions': [extension],
-        'endpoint_sip': {'name': 'iddqd'},
+        'endpoint_sip': {'name': endpoint_name},
     }
     if device:
         line['device_id'] = device['id']
@@ -1204,7 +1208,7 @@ def test_post_update_delete_full_user_no_error(
                 confd.lines(payload['lines'][0]['id']).get().item,
                 has_entries(
                     extensions=contains(has_entries(**extension)),
-                    endpoint_sip=has_entries(name='iddqd'),
+                    endpoint_sip=has_entries(name=line['endpoint_sip']['name']),
                     device_id=device['id'],
                 ),
             )
@@ -1305,7 +1309,7 @@ def test_post_update_delete_full_user_no_error(
             user.pop(
                 'voicemail', None
             )  # The voicemail cannot be updated directly by calling POST /users
-            confd.users(payload['uuid']).put(**user).assert_updated()
+            confd.users(payload['uuid']).put(**user,query_string="recursive=True",).assert_updated()
 
             url = confd.users(payload['uuid'])
 
@@ -1326,9 +1330,14 @@ def test_post_update_delete_full_user_no_error(
                 'unconditional': {'enabled': True, 'destination': '101'},
             }
 
+            payload.pop('call_record_enabled', None)  # Deprecated field
+            payload['lines'][0].pop('caller_id', None)  # cannot set caller id to none
+            payload['lines'][0].pop('caller_id_name', None)  # cannot set caller id to none
+            payload['lines'][0].pop('caller_id_num', None)  # cannot set caller id to none
+
             response = url.put(
                 {
-                    **user_body,
+                    **payload,
                     'fallbacks': {**new_fallbacks},
                     'forwards': {**new_forwards},
                     'groups': [{'uuid': new_group['uuid']}],
@@ -1776,3 +1785,327 @@ def test_post_incalls_existing_extension_wrong_context_type_error(outcall, exten
                 '["Context associated to the extension is not of type \'incall\'"]\n'
             ),
         )
+
+@fixtures.device()
+@fixtures.device()
+def test_update_lines_no_error(
+    device,
+    new_device
+):
+    (
+        _,
+        _,
+        user,
+        _,
+        _,
+        line,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+    ) = generate_user_resources_bodies(
+        context_name=CONTEXT,
+        device=device,
+        endpoint_name='abc'
+    )
+
+    user_body = {
+        'lines': [line],
+        **user,
+    }
+    response = confd.users.post(user_body)
+
+    response.assert_created('users')
+    payload = response.item
+
+    first_line_id = confd.users(payload['uuid']).get().item['lines'][0]['id']
+
+    url = confd.users(payload['uuid'])
+
+    # user update
+    (
+        _,
+        _,
+        _,
+        _,
+        new_extension_to_create,
+        new_line_to_create,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+    ) = generate_user_resources_bodies(
+        context_name=CONTEXT,
+        device=new_device,
+        endpoint_name='def'
+    )
+    payload.pop('call_record_enabled', None)  # Deprecated field
+    response = url.put(
+        {
+            **payload,
+            'lines': [new_line_to_create]
+        },
+        query_string="recursive=True",
+    )
+    response.assert_updated()
+
+    # retrieve the data for the user and check the lines
+    new_line = confd.users(payload['uuid']).get().item['lines'][0]
+    assert_that(new_line['id'], is_not(equal_to(first_line_id)))
+
+    # retrieve the line associated to the user and check its data are correct
+    assert_that(
+        confd.lines(new_line['id']).get().item,
+        has_entries(
+            extensions=contains(has_entries(**new_extension_to_create)),
+            endpoint_sip=has_entries(name=new_line_to_create['endpoint_sip']['name']),
+            device_id=new_device['id'],
+        ),
+    )
+
+    # update the label of the SIP endpoint
+    new_line['endpoint_sip']['label']='new label'
+    response = url.put(
+        {
+            **payload,
+            'lines': [new_line]
+        },
+        query_string="recursive=True",
+    )
+    response.assert_updated()
+
+    # retrieve the data for the user and check the lines
+    assert_that(
+        confd.users(payload['uuid']).get().item,
+        has_entries(
+            lines=contains(has_entries(endpoint_sip=has_entries(label='new label'))),
+        ),
+    )
+
+    # user deletion
+    response = url.delete(recursive=True)
+    response.assert_deleted()
+
+    # verify that user is deleted
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that line is deleted
+    url = confd.lines(new_line['id'])
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that extension is deleted
+    url = confd.extensions(new_line['extensions'][0]['id'])
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that the device is not deleted
+    url = confd.devices(device['id'])
+    response = url.get()
+    response.assert_ok()
+
+
+@fixtures.device()
+def test_update_lines_sip_sccp_error(
+    device,
+):
+    (
+        _,
+        _,
+        user,
+        _,
+        _,
+        line,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+    ) = generate_user_resources_bodies(
+        context_name=CONTEXT,
+        device=device,
+        endpoint_name='abc'
+    )
+
+    user_body = {
+        'lines': [line],
+        **user,
+    }
+    response = confd.users.post(user_body)
+
+    response.assert_created('users')
+    payload = response.item
+
+    # retrieve the user and try to update the user with the same data
+    user = confd.users(payload['uuid']).get().item
+    user.pop('call_record_enabled', None)  # Deprecated field
+    user.pop(
+        'voicemail', None
+    )  # The voicemail cannot be updated directly by calling POST /users
+    confd.users(payload['uuid']).put(**user,query_string="recursive=True",).assert_updated()
+
+    url = confd.users(payload['uuid'])
+
+    # user update
+    del payload['lines'][0]['endpoint_sip']
+    payload['lines'][0]['endpoint_sccp']= {}
+    payload['lines'][0].pop('caller_id', None)  # cannot set caller id to none
+    payload['lines'][0].pop('caller_id_name', None)  # cannot set caller id to none
+    payload['lines'][0].pop('caller_id_num', None)  # cannot set caller id to none
+    response = url.put(
+        {
+            **user,
+            'lines': [payload['lines'][0]]
+        },
+        query_string="recursive=True",
+    )
+    response.assert_status(400)
+    assert_that(
+        response.raw,
+        equal_to(
+            '["There is already an endpoint associated to the line that is not of type SCCP. Cannot update the line with another endpoint type"]\n'
+        ),
+    )
+
+     # user deletion
+    response = url.delete(recursive=True)
+    response.assert_deleted()
+
+    # verify that user is deleted
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that line is deleted
+    url = confd.lines(payload['lines'][0]['id'])
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that extension is deleted
+    url = confd.extensions(payload['lines'][0]['extensions'][0]['id'])
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that the device is not deleted
+    url = confd.devices(device['id'])
+    response = url.get()
+    response.assert_ok()
+
+
+
+@fixtures.device()
+@fixtures.device()
+def test_update_extension_lines_no_error(
+    device,
+    new_device
+):
+    (
+        _,
+        _,
+        user,
+        _,
+        extension,
+        line,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+    ) = generate_user_resources_bodies(
+        context_name=CONTEXT,
+        device=device,
+        endpoint_name='abc'
+    )
+
+    user_body = {
+        'lines': [line],
+        **user,
+    }
+    response = confd.users.post(user_body)
+
+    response.assert_created('users')
+    payload = response.item
+
+    url = confd.users(payload['uuid'])
+
+    # user update
+    (
+        _,
+        _,
+        _,
+        _,
+        new_extension_to_create,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+        _,
+    ) = generate_user_resources_bodies(
+        context_name=CONTEXT,
+        device=new_device,
+        endpoint_name='def'
+    )
+    created_line=payload['lines'][0]
+    created_line['extensions']=[new_extension_to_create]
+
+    payload.pop('call_record_enabled', None)  # Deprecated field
+    payload['lines'][0].pop('caller_id', None)  # cannot set caller id to none
+    payload['lines'][0].pop('caller_id_name', None)  # cannot set caller id to none
+    payload['lines'][0].pop('caller_id_num', None)  # cannot set caller id to none
+    response = url.put(
+        {
+            **payload,
+            'lines': [created_line]
+        },
+        query_string="recursive=True",
+    )
+    response.assert_updated()
+
+    # retrieve the user
+    user = confd.users(payload['uuid']).get().item
+
+    # retrieve the data for the user and check the lines
+    assert_that(
+        confd.users(payload['uuid']).get().item,
+        has_entries(
+            lines=contains(has_entries(extensions=contains(has_entries(**new_extension_to_create)))),
+        ),
+    )
+    assert_that(
+        confd.users(payload['uuid']).get().item,
+        has_entries(
+            lines=not_(contains(
+                has_entries(extensions=contains(has_entries(**extension))))),
+        ),
+    )
+
+     # user deletion
+    response = url.delete(recursive=True)
+    response.assert_deleted()
+
+    # verify that user is deleted
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that line is deleted
+    url = confd.lines(payload['lines'][0]['id'])
+    response = url.get()
+    response.assert_status(404)
+
+    # verify that the device is not deleted
+    url = confd.devices(device['id'])
+    response = url.get()
+    response.assert_ok()