+++ wazo_confd/plugins/extension/middleware.py
@@ -28,3 +28,23 @@ class ExtensionMiddleWare:
     def get_by(self, tenant_uuids, **criteria):
         model = self._service.get_by(tenant_uuids=tenant_uuids, **criteria)
         return self._schema.dump(model)
+
+    def parse_and_update(self, model, body, **kwargs):
+        form = self._schema.load(body, partial=True)
+        updated_fields = self.find_updated_fields(model, form)
+        for name, value in form.items():
+            setattr(model, name, value)
+        self._service.edit(model, updated_fields=updated_fields, **kwargs)
+
+    def find_updated_fields(self, model, form):
+        updated_fields = []
+        for name, value in form.items():
+            try:
+                if getattr(model, name) != value:
+                    updated_fields.append(name)
+            except AttributeError:
+                pass
+
+    def update(self, extension_id, body, tenant_uuids):
+        model = self._service.get(extension_id, tenant_uuids=tenant_uuids)
+        self.parse_and_update(model, body, tenant_uuids=tenant_uuids)