+++ wazo_confd/plugins/user/middleware.py
@@ -8,7 +8,7 @@ from xivo_dao.alchemy.userfeatures import UserFeatures as User
 from xivo_dao.helpers.exception import NotFoundError, ResourceError, InputError
 from xivo_dao.resources.switchboard import dao as switchboard_dao
 
-from .schema import UserListItemSchema, UserSchema
+from .schema import UserListItemSchema, UserSchema, UserListItemSchemaPut
 
 
 class UserMiddleWare:
@@ -18,6 +18,38 @@ class UserMiddleWare:
         self._middleware_handle = middleware_handle
         self._schema = UserListItemSchema()
         self._schema_update = UserSchema()
+        self._schema_update_recursive = UserListItemSchemaPut()
+
+    def create_associate_line(self, user_id, line_body, tenant_uuid, tenant_uuids):
+        line = self._middleware_handle.get('line').create(
+            line_body, tenant_uuid, tenant_uuids
+        )
+        self._middleware_handle.get('user_line_association').associate(
+            user_id, line['id'], tenant_uuids
+        )
+        return line
+
+    def associate_line_device(self,line,device_id, tenant_uuid,tenant_uuids):
+        try:
+            self._middleware_handle.get(
+                'device'
+            ).assign_tenant(device_id, tenant_uuid)
+        except Exception as e:
+            if e is not NotFoundError:
+                raise e
+        self._middleware_handle.get('line_device_association').associate(
+            line['id'], device_id, tenant_uuid, tenant_uuids
+        )
+
+    def create_line(self, user_id, line_body, tenant_uuid, tenant_uuids):
+        line = self.create_associate_line(user_id, line_body, tenant_uuid,
+                                          tenant_uuids)
+
+        device_id = line_body.get('device_id', None)
+        if device_id:
+            self.associate_line_device(line, device_id, tenant_uuid, tenant_uuids)
+            line['device_id'] = device_id
+        return line
 
     def create(self, body, tenant_uuid, tenant_uuids):
         forwards = body.pop('forwards', None) or []
@@ -67,27 +99,9 @@ class UserMiddleWare:
 
         if lines:
             for line_body in lines:
-                line = self._middleware_handle.get('line').create(
-                    line_body, tenant_uuid, tenant_uuids
-                )
-                self._middleware_handle.get('user_line_association').associate(
-                    user_dict['id'], line['id'], tenant_uuids
-                )
+                line=self.create_line(user_dict['id'], line_body, tenant_uuid,
+                                                  tenant_uuids)
                 user_dict['lines'].append(line)
-
-                device_id = line_body.get('device_id', None)
-                if device_id:
-                    try:
-                        self._middleware_handle.get('device').assign_tenant(
-                            device_id, tenant_uuid
-                        )
-                    except Exception as e:
-                        if e is not NotFoundError:
-                            raise e
-                    self._middleware_handle.get('line_device_association').associate(
-                        line['id'], device_id, tenant_uuid, tenant_uuids
-                    )
-                    line['device_id'] = device_id
 
             Session.expire(model, ['user_lines'])
 
@@ -243,6 +257,20 @@ class UserMiddleWare:
 
         return user_dict
 
+    def delete_line(self,device_id, line_id, user_id, tenant_uuid, tenant_uuids):
+        if device_id:
+            self._middleware_handle.get(
+                'device'
+            ).reset_autoprov(device_id, tenant_uuid)
+
+        # process the line itself
+        self._middleware_handle.get('user_line_association').dissociate(
+            user_id, line_id, tenant_uuids
+        )
+        self._middleware_handle.get('line').delete(
+            line_id, tenant_uuid, tenant_uuids, recursive=True
+        )
+
     def delete(self, user_id, tenant_uuid, tenant_uuids, recursive=False):
         user = self._service.get(user_id, tenant_uuids=tenant_uuids)
         if not recursive:
@@ -268,19 +296,7 @@ class UserMiddleWare:
 
             for line in user.lines:
                 # process the device associated to the line
-                device_id = line.device
-                if device_id:
-                    self._middleware_handle.get('device').reset_autoprov(
-                        device_id, tenant_uuid
-                    )
-
-                # process the line itself
-                self._middleware_handle.get('user_line_association').dissociate(
-                    user.uuid, line.id, tenant_uuids
-                )
-                self._middleware_handle.get('line').delete(
-                    line.id, tenant_uuids, recursive=True
-                )
+                self.delete_line(line.device, line.id, user.uuid, tenant_uuid, tenant_uuids)
 
                 Session.expire(user, ['user_lines'])
 
@@ -329,7 +345,7 @@ class UserMiddleWare:
                 pass
         return updated_fields
 
-    def update(self, user_id, body, tenant_uuids, recursive=False):
+    def update(self, user_id, body, tenant_uuid, tenant_uuids, recursive=False):
         user = self._service.get(user_id, tenant_uuids=tenant_uuids)
         if not recursive:
             self.parse_and_update(user, body)
@@ -337,9 +353,10 @@ class UserMiddleWare:
             fallbacks = body.pop('fallbacks', None) or []
             forwards = body.pop('forwards', None) or []
 
-            form = self._schema.load(body)
+            form = self._schema_update_recursive.load(body)
 
             groups = form.pop('groups', None) or []
+            lines = form.pop('lines', None) or []
 
             if fallbacks:
                 self._middleware_handle.get('user_fallback_association').associate(
@@ -358,3 +375,27 @@ class UserMiddleWare:
             self._middleware_handle.get('user_group_association').associate_all_groups(
                 {'groups': groups}, user_id
             )
+
+            # lines
+            user = self._service.get(user_id, tenant_uuids=tenant_uuids)
+            existing_lines= {el.id: el for el in user.lines}
+
+            for line_body in lines:
+                device_id = line_body.get('device_id', None)
+                if 'id' in line_body and line_body['id'] in existing_lines:
+                    old_device_id=existing_lines[line_body['id']].device_id
+                    self._middleware_handle.get('line').update(line_body['id'], line_body, tenant_uuid, tenant_uuids)
+                    #if device_id not the same, so we must dissociate the old one and associate the new line
+                    if device_id!=old_device_id:
+                        self._middleware_handle.get(
+                        'device'
+                    ).reset_autoprov(old_device_id, tenant_uuid)
+                    self.associate_line_device({'id':line_body['id']}, device_id, tenant_uuid, tenant_uuids)
+
+                    del existing_lines[line_body['id']]
+                else:
+                    line = self.create_associate_line(user_id, line_body,tenant_uuid, tenant_uuids)
+                    if device_id:
+                        self.associate_line_device(line, device_id, tenant_uuid, tenant_uuids)
+            for line_id in existing_lines:
+                self.delete_line(existing_lines[line_id].device, line_id, user.uuid, tenant_uuid, tenant_uuids)