+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 23.03
 
 * POST on `/users?recursive=true` updated, to provide a way to create/reuse incalls.
-* PUT on `/users?recursive=true` added, to provide a way to update groups for a specific user.
+* PUT on `/users?recursive=true` added, to provide a way to update groups and lines for a specific user.
 
 ## 23.02